MAIN/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-crispy-forms==2.1
django-etc==1.4.0
django-hitcount==1.3.5
django-resized==1.0.2
django-taggit==5.0.1
django-tinymce==4.0.0
pillow==10.3.0
sqlparse==0.5.0
termcolor==2.4.0
termcolors==0.1.0
tzdata==2024.1
